# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Missing permission"
          }
        },
        "summary": "Get users by id",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Missing permission"
          }
        },
        "summary": "Update users",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Missing permission"
          }
        },
        "summary": "Soft-delete users",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/user-payment-methods": {
      "get": {
        "operationId": "UserPaymentMethodsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter by owner user id (staff only)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Missing permission or access denied"
          }
        },
        "summary": "List user payment methods (scoped to current user unless staff)",
        "tags": [
          "user-payment-methods"
        ]
      },
      "post": {
        "operationId": "UserPaymentMethodsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserPaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "403": {
            "description": "Missing permission or access denied"
          }
        },
        "summary": "Create user payment method",
        "tags": [
          "user-payment-methods"
        ]
      }
    },
    "/api/user-payment-methods/{id}": {
      "get": {
        "operationId": "UserPaymentMethodsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Missing permission or access denied"
          }
        },
        "summary": "Get user payment method by id",
        "tags": [
          "user-payment-methods"
        ]
      },
      "patch": {
        "operationId": "UserPaymentMethodsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserPaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Missing permission or access denied"
          }
        },
        "summary": "Update user payment method",
        "tags": [
          "user-payment-methods"
        ]
      },
      "delete": {
        "operationId": "UserPaymentMethodsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Missing permission or access denied"
          }
        },
        "summary": "Soft-delete user payment method",
        "tags": [
          "user-payment-methods"
        ]
      }
    },
...
    "/api/payments/{id}": {
      "get": {
        "operationId": "PaymentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get payments by id",
        "tags": [
          "payments"
        ]
      },
      "patch": {
        "operationId": "PaymentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update payments",
        "tags": [
          "payments"
        ]
      },
      "delete": {
        "operationId": "PaymentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Soft-delete payments",
        "tags": [
          "payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Africa Tourism Gate API",
    "description": "REST API aligned with africatourismgate MySQL schema",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserPaymentMethodDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "card",
              "paypal",
              "other"
            ]
          },
          "provider": {
            "type": "string",
            "example": "visa"
          },
          "lastFour": {
            "type": "string",
            "example": "4242"
          },
          "externalToken": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean",
            "default": false
          },
          "userId": {
            "type": "string",
            "description": "Staff only: target user id"
          }
        },
        "required": [
          "type"
        ]
      },
      "UpdateUserPaymentMethodDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "card",
              "paypal",
              "other"
            ]
          },
          "provider": {
            "type": "object"
          },
          "lastFour": {
            "type": "object"
          },
          "externalToken": {
            "type": "object"
          },
          "isDefault": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}